Decompose TerminalChat into memoized regions with grouped state
index.tsx drops from 1,223 to 153 lines: state moves into focused hooks
under ui/state/ (useReducer for co-updating groups), JSX splits into
four React.memo regions (transcript, status, input, modal host), and
keystroke-local state lives inside InputRegion so typing re-renders
only the input. The resetKey full-remount is replaced by an in-place
resetSession(); modals.tsx splits per modal; chat-input's key handler
extracts to chat-input-keys.ts. Render isolation is pinned by
tests/ui-render-isolation.test.ts using ink-testing-library.

ref #184
