+++ package.json
@@ -58,6 +58,7 @@
     "@types/node": "^22.10.2",
     "@types/react": "^19.2.7",
     "@vitest/coverage-v8": "^4.0.16",
+    "ink-testing-library": "^4.0.0",
     "sharp": "^0.34.5",
     "typescript": "^5.7.2",
     "vitest": "^4.0.16"
