+++ src/ui/components.tsx
@@ -2,13 +2,19 @@
  * UI Module - Utility Components
  *
  * Animated thinking displays, state transitions, streaming indicators.
- * Colors sourced from active palette, animations from active skin.
+ * Colors sourced from active palette, animations from active skin. Pure frame
+ * generators and the transition hook live in animations.ts.
  */
 
 import React, { useState, useEffect, useRef } from 'react';
 import { Box, Text, useStdout } from 'ink';
 import { getCurrentSkin, getSpinnerFrames, getInkColor } from '../hud/api.js';
 import type { ThinkingState, ActivityState } from './types.js';
+import {
+  waveFrame, neuralFrame, circuitFrame, dnaFrame, pulseBarFrame, cascadeFrame,
+  scanLine, getAnimationStyle, useTransition, transitionWipe, formatElapsed,
+  elapsedDots, ORBIT_CHARS,
+} from './animations.js';
 
 // ============================================================================
 // Constants & Helpers
@@ -37,214 +43,6 @@ export function getStateSpinner(state: 'thinking' | 'processing' | 'streaming'):
 export const DEFAULT_SPINNER = ['⠋', '⠙', '⠹', '⠸', '⠼', '⠴', '⠦', '⠧', '⠇', '⠏'];
 
 // ============================================================================
-// Thinking Animation Patterns
-// ============================================================================
-
-type AnimationStyle = 'wave' | 'neural' | 'circuit' | 'dna' | 'pulse-bar' | 'orbit' | 'cascade' | 'minimal';
-
-/** Characters for each animation style */
-const WAVE_CHARS = '  ░▒▓█▓▒░  ';
-const NEURAL_NODES = ['◇', '◆', '○', '●', '◎', '◉'];
-const CIRCUIT_CHARS = ['─', '┐', '│', '└', '─', '┌', '│', '┘'];
-const DNA_LEFT =  ['╭', '│', '╰', ' ', ' ', '╭', '│', '╰'];
-const DNA_RIGHT = [' ', ' ', '╭', '│', '╰', ' ', ' ', '╭'];
-const DNA_BASES = ['A', 'T', 'G', 'C'];
-const ORBIT_CHARS = [
-  '    ◠    ',
-  '  ◜   ◝  ',
-  ' ◜  ●  ◝ ',
-  '  ◟   ◞  ',
-  '    ◡    ',
-];
-
-/** Generate a wave animation line */
-function waveFrame(tick: number, width: number): string {
-  let line = '';
-  for (let x = 0; x < width; x++) {
-    const phase = (x * 0.3 + tick * 0.4) % WAVE_CHARS.length;
-    line += WAVE_CHARS[Math.floor(phase)] || ' ';
-  }
-  return line;
-}
-
-/** Generate neural network activity */
-function neuralFrame(tick: number, width: number): string {
-  let line = '';
-  for (let x = 0; x < width; x++) {
-    if (x % 4 === 0) {
-      const nodeIdx = (tick + x) % NEURAL_NODES.length;
-      line += NEURAL_NODES[nodeIdx];
-    } else if (x % 4 === 2) {
-      const active = ((tick + x * 3) % 7) < 3;
-      line += active ? '─' : ' ';
-    } else {
-      line += ' ';
-    }
-  }
-  return line;
-}
-
-/** Generate circuit trace pattern */
-function circuitFrame(tick: number, width: number): string {
-  let line = '';
-  const pulsePos = tick % width;
-  for (let x = 0; x < width; x++) {
-    const charIdx = (x + tick) % CIRCUIT_CHARS.length;
-    const dist = Math.abs(x - pulsePos);
-    if (dist < 3) {
-      line += CIRCUIT_CHARS[charIdx];
-    } else if (x % 5 === 0) {
-      line += '·';
-    } else {
-      line += ' ';
-    }
-  }
-  return line;
-}
-
-/** Generate DNA helix pattern (2 lines) */
-function dnaFrame(tick: number, _width: number): string[] {
-  const len = 24;
-  let top = ' ';
-  let bot = ' ';
-  for (let x = 0; x < len; x++) {
-    const phase = (x + tick) % 8;
-    const lt = DNA_LEFT[phase];
-    const rt = DNA_RIGHT[phase];
-    if (lt !== ' ' && rt !== ' ') {
-      // Crossover — show a base pair
-      const base = DNA_BASES[(x + tick) % DNA_BASES.length];
-      top += lt + base;
-      bot += rt + base;
-    } else {
-      top += lt + ' ';
-      bot += rt + ' ';
-    }
-  }
-  return [top, bot];
-}
-
-/** Generate an animated pulse bar */
-function pulseBarFrame(tick: number, width: number): string {
-  const barWidth = Math.min(width - 4, 32);
-  let bar = '│';
-  for (let x = 0; x < barWidth; x++) {
-    const intensity = Math.sin((x / barWidth) * Math.PI + tick * 0.3);
-    if (intensity > 0.8) bar += '█';
-    else if (intensity > 0.5) bar += '▓';
-    else if (intensity > 0.2) bar += '▒';
-    else if (intensity > -0.2) bar += '░';
-    else bar += ' ';
-  }
-  bar += '│';
-  return bar;
-}
-
-/** Generate cascade / waterfall pattern */
-function cascadeFrame(tick: number, width: number): string[] {
-  const lines: string[] = [];
-  const cascadeWidth = Math.min(width - 2, 36);
-  for (let row = 0; row < 3; row++) {
-    let line = '';
-    for (let x = 0; x < cascadeWidth; x++) {
-      const drop = ((x * 7 + row * 13 + tick * 3) % 23);
-      if (drop < 2) line += '│';
-      else if (drop < 4) line += '┊';
-      else if (drop < 6) line += '·';
-      else line += ' ';
-    }
-    lines.push(line);
-  }
-  return lines;
-}
-
-/** Pick animation style from skin config or infer from skin name */
-function getAnimationStyle(): AnimationStyle {
-  const skin = getCurrentSkin();
-  // Explicit skin preference wins
-  if (skin.animations?.thinkingStyle) return skin.animations.thinkingStyle;
-
-  const skinName = skin.name.toLowerCase();
-  // Infer from skin name / vibe
-  if (skinName.includes('matrix') || skinName.includes('neo') || skinName.includes('hack'))
-    return 'cascade';
-  if (skinName.includes('retro') || skinName.includes('basic') || skinName.includes('arcade'))
-    return 'circuit';
-  if (skinName.includes('sci') || skinName.includes('trek') || skinName.includes('space'))
-    return 'neural';
-  if (skinName.includes('bio') || skinName.includes('nature') || skinName.includes('organic'))
-    return 'dna';
-  if (skinName.includes('minimal') || skinName.includes('clean'))
-    return 'minimal';
-  return 'wave';
-}
-
-// ============================================================================
-// State Transition System
-// ============================================================================
-
-type TransitionPhase = 'enter' | 'active' | 'exit';
-
-function useTransition(isActive: boolean, enterMs = 300, exitMs = 200): { phase: TransitionPhase; progress: number } {
-  const [phase, setPhase] = useState<TransitionPhase>(isActive ? 'active' : 'exit');
-  const [progress, setProgress] = useState(isActive ? 1 : 0);
-  const startTime = useRef(Date.now());
-
-  useEffect(() => {
-    if (isActive && phase !== 'active') {
-      setPhase('enter');
-      startTime.current = Date.now();
-    } else if (!isActive && phase === 'active') {
-      setPhase('exit');
-      startTime.current = Date.now();
-    }
-  }, [isActive]);
-
-  useEffect(() => {
-    if (phase === 'enter' || phase === 'exit') {
-      const duration = phase === 'enter' ? enterMs : exitMs;
-      const timer = setInterval(() => {
-        const elapsed = Date.now() - startTime.current;
-        const p = Math.min(1, elapsed / duration);
-        setProgress(phase === 'enter' ? p : 1 - p);
-        if (p >= 1) {
-          setPhase(phase === 'enter' ? 'active' : 'exit');
-          setProgress(phase === 'enter' ? 1 : 0);
-          clearInterval(timer);
-        }
-      }, 30);
-      return () => clearInterval(timer);
-    }
-  }, [phase]);
-
-  return { phase, progress };
-}
-
-/** Build a transition-in wipe line */
-function transitionWipe(text: string, progress: number): string {
-  const visibleChars = Math.floor(text.length * progress);
-  return text.slice(0, visibleChars) + ' '.repeat(Math.max(0, text.length - visibleChars));
-}
-
-// ============================================================================
-// Elapsed Time Visualization
-// ============================================================================
-
-function formatElapsed(seconds: number): string {
-  if (seconds < 60) return `${seconds}s`;
-  const m = Math.floor(seconds / 60);
-  const s = seconds % 60;
-  return `${m}m${s.toString().padStart(2, '0')}s`;
-}
-
-/** Tiny progress dots that grow with time */
-function elapsedDots(seconds: number): string {
-  const filled = Math.min(seconds, 20);
-  const dots = '·'.repeat(Math.max(0, 20 - filled)) + '•'.repeat(Math.min(filled, 10));
-  return dots.slice(0, 20);
-}
-
-// ============================================================================
 // Components
 // ============================================================================
 
@@ -255,7 +53,7 @@ export function Separator() {
   return <Text color={borderColor} dimColor>{'─'.repeat(width)}</Text>;
 }
 
-export function ThinkingDisplay({ state }: { state: ThinkingState }) {
+function ThinkingDisplayInner({ state }: { state: ThinkingState }) {
   const [frame, setFrame] = useState(0);
   const [elapsed, setElapsed] = useState(0);
   const startTime = useRef(Date.now());
@@ -357,7 +155,11 @@ export function ThinkingDisplay({ state }: { state: ThinkingState }) {
   );
 }
 
-export function ProcessingIndicator({ label }: { label: string }) {
+// Memoized so unrelated parent re-renders (streaming chunks, stats ticks) don't
+// re-run the animation build between its own timer frames.
+export const ThinkingDisplay = React.memo(ThinkingDisplayInner);
+
+function ProcessingIndicatorInner({ label }: { label: string }) {
   const [frame, setFrame] = useState(0);
   const spinFrames = getStateSpinner('processing');
   const primaryColor = getInkColor('primary');
@@ -387,24 +189,10 @@ export function ProcessingIndicator({ label }: { label: string }) {
   );
 }
 
-/** A subtle scanning line effect */
-function scanLine(tick: number, width: number): string {
-  const pos = tick % (width * 2);
-  const actual = pos < width ? pos : width * 2 - pos;
-  let line = '';
-  for (let x = 0; x < width; x++) {
-    const dist = Math.abs(x - actual);
-    if (dist === 0) line += '█';
-    else if (dist === 1) line += '▓';
-    else if (dist === 2) line += '▒';
-    else if (dist === 3) line += '░';
-    else line += '·';
-  }
-  return line;
-}
+export const ProcessingIndicator = React.memo(ProcessingIndicatorInner);
 
 /** Brief splash overlay shown when switching themes (auto-dismisses) */
-export function SplashOverlay({ art, color, onDone }: { art: string[]; color: string; onDone: () => void }) {
+function SplashOverlayInner({ art, color, onDone }: { art: string[]; color: string; onDone: () => void }) {
   const [visibleLines, setVisibleLines] = useState(0);
 
   useEffect(() => {
@@ -427,7 +215,9 @@ export function SplashOverlay({ art, color, onDone }: { art: string[]; color: st
   );
 }
 
-export function StreamingIndicator({ activity }: { activity?: ActivityState }) {
+export const SplashOverlay = React.memo(SplashOverlayInner);
+
+function StreamingIndicatorInner({ activity }: { activity?: ActivityState }) {
   const [frame, setFrame] = useState(0);
   const [elapsed, setElapsed] = useState(0);
   const pulseFrames = getStateSpinner('streaming');
@@ -480,6 +270,8 @@ export function StreamingIndicator({ activity }: { activity?: ActivityState }) {
   );
 }
 
+export const StreamingIndicator = React.memo(StreamingIndicatorInner);
+
 // ============================================================================
 // State Transition Overlay
 // ============================================================================
@@ -488,7 +280,7 @@ export function StreamingIndicator({ activity }: { activity?: ActivityState }) {
  * Renders a brief transition effect between processing states.
  * Mount when transitioning, unmount after onComplete fires.
  */
-export function StateTransition({ from, to, onComplete }: {
+function StateTransitionInner({ from, to, onComplete }: {
   from: 'idle' | 'thinking' | 'streaming' | 'done';
   to: 'idle' | 'thinking' | 'streaming' | 'done';
   onComplete: () => void;
@@ -566,3 +358,5 @@ export function StateTransition({ from, to, onComplete }: {
     </Box>
   );
 }
+
+export const StateTransition = React.memo(StateTransitionInner);
